Rework how plugins are updated
We used to clone the repo and do a full `npm install` and `grunt setup`
so we could run all the `wp plugin update` commands locally which was
slow and fragile.

Now we run the `wp plugin update` commands on the remote WPE instance
and then rsync any changed files down from production.